requirements.txt
fastapi 
uvicorn
onnxruntime
pillow
tf2onnx

